Move collada_element! to macros module
